==> data_analysis/01-3.ipynb <==
import gdown

gdown.download('https://bit.ly/3eecMKZ',
               '남산도서관 장서 대출목록 (2021년 04월).csv', quiet=False)

with open('남산도서관 장서 대출목록 (2021년 04월).csv') as f:
  print(f.readline)

import chardet

with open('남산도서관 장서 대출목록 (2021년 04월).csv', mode='rb') as f:
  d = f.readline()
print(chardet.detect(d))

with open('남산도서관 장서 대출목록 (2021년 04월).csv', encoding='EUC-KR') as f:
  print(f.readline())
  print(f.readline())

import pandas as pd

df = pd.read_csv('남산도서관 장서 대출목록 (2021년 04월).csv', encoding='EUC-KR')

df = pd.read_csv('남산도서관 장서 대출목록 (2021년 04월).csv', encoding='euc-kr', dtype={'ISBN':str, '세트 ISBN':str, '주제분류번호':str})

df.head()

df.to_csv('ns_202104.csv')

with open('ns_202104.csv') as f:
  for i in range(3):
    print(f.readline(), end='')

ns_df = pd.read_csv('ns_202104.csv', low_memory=False)
ns_df.head()

df.to_csv('ns_202104.csv', index=False)
