==> data_analysis/04-1.ipynb <==
import gdown
gdown.download('https://bit.ly/3736JW1', 'ns_book6.csv', quiet=False)

import pandas as pd
ns_book6 = pd.read_csv('ns_book6.csv', low_memory=False)
ns_book6.head()

ns_book6.describe()

sum(ns_book6['도서권수'] == 0)

ns_book7 = ns_book6[ns_book6['도서권수'] > 0]

ns_book7.describe()

ns_book7.describe(include='object')

x = [10, 20, 30]
sum = 0
for i in x:
    sum += i
print("평균:", sum / len(x))

ns_book7['대출건수'].mean()

ns_book7['대출건수'].median()


temp_df = pd.DataFrame([1,2,3,4])
temp_df.median()

ns_book7['대출건수'].drop_duplicates().median()

ns_book7['대출건수'].min()

ns_book7['대출건수'].max()

ns_book7['대출건수'].quantile(0.25)

ns_book7['대출건수'].quantile([0.25,0.5,0.75])

pd.Series([1,2,3,4,5]).quantile(0.9)

borrow_10_flag = ns_book7['대출건수'] < 10

borrow_10_flag.sum()

borrow_10_flag.mean()

ns_book7['대출건수'].quantile(0.65)

ns_book7['대출건수'].var()

ns_book7['대출건수'].std()

import numpy as np

diff = ns_book7['대출건수'] - ns_book7['대출건수'].mean()
np.sqrt(np.sum(diff**2) / (len(ns_book7)-1))

ns_book7['도서명'].mode()

ns_book7['발행년도'].mode()

ns_book7.mean(numeric_only=True)

ns_book7.loc[:, '도서명':].mode()

ns_book7.to_csv('ns_book7.csv', index=False)

import numpy as np
np.mean(ns_book7['대출건수'])

np.average(ns_book7['대출건수'], weights=1/ns_book7['도서권수'])

np.mean(ns_book7['대출건수']/ns_book7['도서권수'])

ns_book7['대출건수'].sum()/ns_book7['도서권수'].sum()

np.median(ns_book7['대출건수'])

np.min(ns_book7['대출건수'])

np.max(ns_book7['대출건수'])

ns_book7['대출건수'].var()

np.var(ns_book7['대출건수'])

ns_book7['대출건수'].var(ddof=0)

np.var(ns_book7['대출건수'], ddof=1)

values, counts = np.unique(ns_book7['도서명'], return_counts=True)
max_idx = np.argmax(counts)
values[max_idx]

values[np.argmax(counts)]
